requirements.txt
llama-cpp-python>=0.2.80
huggingface-hub>=0.23.0
fastapi>=0.111.0
uvicorn>=0.30.0
click>=8.1.0
rich>=13.7.0
